+++ Firestore/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
+- [fixed] Fix FAILED_PRECONDITION when writing to a deleted document in a
+  transaction (#10431).
 - [fixed] Fixed data race in credentials provider (#10393).
 
 # 10.0.0